deeppavlov/requirements/intent_catcher.txt
xeger==0.3.5
numpy==1.18.0
scikit-learn==0.21.2
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
tensorflow-hub==0.10.0
